requirements.txt
python-telegram-bot>=22.0,<23.0
aiohttp>=3.9,<4.0
mistune>=3.0,<4.0
pytest>=8.0
pytest-asyncio>=0.24
pytest-cov>=6.0
